Move silx.gui.test.utils to silx.gui.testutils
This avoids issue with import
