+++ silx/gui/testutils.py
